added routes and forms for shop address, NOT COMPLETED

+++ src/Form/Profile/DeleteAddressForm.php
@@ -4,6 +4,7 @@ namespace App\Form\Profile;
 
 use App\Lib\Form\ABaseForm;
 use App\Repository\AddressRepository;
+use App\Repository\UserRepository;
 use Doctrine\ORM\EntityNotFoundException;
 use Symfony\Component\HttpFoundation\Exception\BadRequestException;
 use Symfony\Component\HttpFoundation\Request;
@@ -12,7 +13,8 @@ use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 class DeleteAddressForm extends ABaseForm
 {
     public function __construct(
-        private readonly AddressRepository $addressRepository
+        private readonly AddressRepository $addressRepository,
+        private readonly UserRepository $userRepository
     ) {
 
     }
@@ -34,14 +36,22 @@ class DeleteAddressForm extends ABaseForm
     public function execute(Request $request): String
     {
         $form = self::getParams($request);
+        $user_id = $form["route"]["user_id"];
         $addressId = $form["route"]["address_id"];
         $address = $this->addressRepository->find($addressId);
+        $user = $this->userRepository->find($user_id);
 
         if (!$addressId) {
             throw new BadRequestHttpException("invalid address id");
         }
 
+        if (!$user) {
+            throw new BadRequestHttpException("invalid user id");
+        }
+
+        $user->removeAddress($address);
         $this->addressRepository->remove($address);
+        $this->userRepository->flush();
         $this->addressRepository->flush();
 
         return "address removed";
